Making corrections/modifications
